Apps/requirements/prod-LangChain.txt
langchain >= 0.2.1
langchain-core >= 0.2.3
langchain-community >= 0.2.1
langchain-openai >= 0.1.8 # == 0.0.5 (crewai), >= 0.1.1
openai >= 1.30.5
tiktoken >= 0.7.0 # >=0.5.2 | >= 0.6.0
langchain-anthropic >= 0.1.15
langchain-qdrant >= 0.1.0
qdrant-client >= 1.9.1
langchain_cohere >= 0.1.5
cohere >= 5.5.4
langchain-groq >= 0.1.4
groq >= 0.8.0
langgraph >= 0.0.61

langchainhub >= 0.1.15
langsmith >= 0.1.57
pydantic == 2.6.1 # 2.6.1 1.10.9 1.10.13

lark

# Retrievers
rank-bm25

# Store
chromadb >= 0.5.0
faiss-cpu >= 1.8.0

# Tools
tools
google-search-results
wikipedia
duckduckgo-search

# Document Loaders
youtube-transcript-api
docx2txt
pytube
unstructured
unstructured_inference
pytesseract
pdf2image
pdfminer.six
pillow_heif
opencv-python
pytesseract
pikepdf

beautifulsoup4

#===============================================================================

boto3 >= 1.34.117
context-python
faker
fastapi >= 0.111.0
uvicorn
httpx >= 0.27.0
ipykernel
loguru
openpyxl
psycopg2-binary >= 2.9.9 # Postgresql
python-dateutil
python-dotenv
requests
streamlit >= 1.35.0
streamlit_feedback
tqdm
# transformers
urllib3
watchdog